removed useless class, reorganization
