+++ src/main/java/tkcy/simpleaddon/api/recipes/logic/IExtraRecipeLogic.java
@@ -2,8 +2,6 @@ package tkcy.simpleaddon.api.recipes.logic;
 
 import org.jetbrains.annotations.NotNull;
 
-import tkcy.simpleaddon.api.recipes.logic.newway.IRecipeLogicContainer;
-
 public interface IExtraRecipeLogic {
 
     @NotNull
