+++ src/main/java/tkcy/simpleaddon/api/recipes/logic/impl/RecipeLogicsContainer.java
@@ -1,4 +1,4 @@
-package tkcy.simpleaddon.api.recipes.logic.newway;
+package tkcy.simpleaddon.api.recipes.logic.impl;
 
 import java.util.*;
 
@@ -16,6 +16,9 @@ import gregtech.api.recipes.Recipe;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
+import tkcy.simpleaddon.api.recipes.logic.IRecipeLogicContainer;
+import tkcy.simpleaddon.api.recipes.logic.IRecipePropertiesValueMap;
+import tkcy.simpleaddon.api.recipes.logic.RecipeLogicType;
 import tkcy.simpleaddon.api.recipes.properties.IRecipePropertyHelper;
 
 @Getter
