+++ src/main/java/tkcy/simpleaddon/common/metatileentities/primitive/AnvilMetatileEntity.java
@@ -36,12 +36,12 @@ import codechicken.lib.vec.Matrix4;
 import tkcy.simpleaddon.api.machines.IRightClickItemTransfer;
 import tkcy.simpleaddon.api.machines.IUnificationToolMachine;
 import tkcy.simpleaddon.api.machines.ToolLogicMetaTileEntity;
+import tkcy.simpleaddon.api.recipes.logic.IRecipeLogicContainer;
 import tkcy.simpleaddon.api.recipes.logic.IToolRecipeLogic;
 import tkcy.simpleaddon.api.recipes.logic.OnBlockRecipeLogic;
-import tkcy.simpleaddon.api.recipes.logic.newway.IRecipeLogicContainer;
-import tkcy.simpleaddon.api.recipes.logic.newway.InWorldRecipeLogic;
-import tkcy.simpleaddon.api.recipes.logic.newway.RecipeLogicsContainer;
-import tkcy.simpleaddon.api.recipes.logic.newway.ToolLogic;
+import tkcy.simpleaddon.api.recipes.logic.impl.InWorldRecipeLogic;
+import tkcy.simpleaddon.api.recipes.logic.impl.RecipeLogicsContainer;
+import tkcy.simpleaddon.api.recipes.logic.impl.ToolLogic;
 import tkcy.simpleaddon.api.recipes.recipemaps.TKCYSARecipeMaps;
 import tkcy.simpleaddon.modules.toolmodule.ToolsModule;
