+++ src/main/java/tkcy/simpleaddon/common/metatileentities/primitive/MetaTileEntityWoodWorkshop.java
@@ -32,8 +32,10 @@ import codechicken.lib.render.pipeline.ColourMultiplier;
 import codechicken.lib.render.pipeline.IVertexOperation;
 import codechicken.lib.vec.Matrix4;
 import tkcy.simpleaddon.api.machines.*;
-import tkcy.simpleaddon.api.recipes.logic.OnBlockRecipeLogic;
-import tkcy.simpleaddon.api.recipes.logic.newway.*;
+import tkcy.simpleaddon.api.recipes.logic.*;
+import tkcy.simpleaddon.api.recipes.logic.impl.InWorldRecipeLogic;
+import tkcy.simpleaddon.api.recipes.logic.impl.RecipeLogicsContainer;
+import tkcy.simpleaddon.api.recipes.logic.impl.ToolLogic;
 import tkcy.simpleaddon.api.recipes.recipemaps.TKCYSARecipeMaps;
 import tkcy.simpleaddon.api.unification.ore.TKCYSAOrePrefix;
 import tkcy.simpleaddon.modules.toolmodule.ToolsModule;
